percpu: add feature "sp-naive" to pass unittest
