+++ crates/percpu_macros/Cargo.toml
@@ -6,6 +6,12 @@ authors = ["Yuekai Jia <equation618@gmail.com>"]
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
+[features]
+# For single CPU use, just make the per-CPU data a global variable.
+sp-naive = []
+
+default = []
+
 [dependencies]
 proc-macro2 = "1.0"
 quote = "1.0"
