+++ modules/axhal/Cargo.toml
@@ -48,3 +48,6 @@ arm_gic = { path = "../../crates/arm_gic" }
 
 [build-dependencies]
 axconfig = { path = "../axconfig" }
+
+[dev-dependencies]
+percpu = { path = "../../crates/percpu", features = ["sp-naive"] }
