+++ modules/axruntime/Cargo.toml
@@ -30,3 +30,6 @@ axlog = { path = "../axlog" }
 axnet = { path = "../axnet", optional = true }
 axdisplay = { path = "../axdisplay", optional = true }
 axtask = { path = "../axtask", default-features = false }
+
+[dev-dependencies]
+percpu = { path = "../../crates/percpu", features = ["sp-naive"] }
